+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/DefaultProjectSnapshotManagerAccessor.cs
@@ -11,7 +11,6 @@ using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Razor;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 using Microsoft.Extensions.Options;
-using Microsoft.CodeAnalysis.ExternalAccess.Razor.Remote;
 using Microsoft.AspNetCore.Razor.LanguageServer.Common;
 using Microsoft.AspNetCore.Razor.LanguageServer.EndpointContracts;
 using Microsoft.VisualStudio.LanguageServer.Protocol;
@@ -19,73 +18,6 @@ using Microsoft.VisualStudio.RpcContracts.Documents;
 using System.Text;
 
 namespace Microsoft.AspNetCore.Razor.LanguageServer;
-
-internal class LspRazorGeneratedDocumentProvider : IRazorGeneratedDocumentProvider
-{
-    readonly ClientNotifierServiceBase _notifier;
-
-    public LspRazorGeneratedDocumentProvider(ClientNotifierServiceBase notifier)
-    {
-        _notifier = notifier;
-    }
-
-    public async Task<(string CSharp, string Html, string Json)> GetGeneratedDocumentAsync(IDocumentSnapshot documentSnapshot)
-    {
-        var projectRoot = documentSnapshot.Project.FilePath.Substring(0, documentSnapshot.Project.FilePath.LastIndexOf("/"));
-        var documentName = GetIdentifierFromPath(documentSnapshot.FilePath?.Substring(projectRoot.Length + 1) ?? "");
-
-        var csharp = await RequestOutput(documentName + ".rsg.cs");
-        var html = await RequestOutput(documentName + ".rsg.html");
-        var json = await RequestOutput(documentName + ".rsg.json");
-
-
-        return (csharp, html, json);
-
-        async Task<string> RequestOutput(string name)
-        {
-            var request = new HostOutputRequest()
-            {
-                TextDocument = new TextDocumentIdentifier()
-                {
-                    Uri = new UriBuilder()
-                    {
-                        Scheme = Uri.UriSchemeFile,
-                        Path = documentSnapshot.FilePath,
-                        Host = string.Empty,
-                    }.Uri
-                },
-                GeneratorName = "Microsoft.NET.Sdk.Razor.SourceGenerators.RazorSourceGenerator",
-                RequestedOutput = name,
-            };
-
-            var response = await _notifier.SendRequestAsync<HostOutputRequest, HostOutputResponse>(RazorLanguageServerCustomMessageTargets.RazorHostOutputsEndpointName, request, CancellationToken.None);
-            return response.Output ?? string.Empty;
-        }
-    }
-
-    //copied from the generator
-    internal static string GetIdentifierFromPath(string filePath)
-    {
-        var builder = new StringBuilder(filePath.Length);
-
-        for (var i = 0; i < filePath.Length; i++)
-        {
-            switch (filePath[i])
-            {
-                case ':' or '\\' or '/':
-                case char ch when !char.IsLetterOrDigit(ch):
-                    builder.Append('_');
-                    break;
-                default:
-                    builder.Append(filePath[i]);
-                    break;
-            }
-        }
-
-        return builder.ToString();
-    }
-
-}
 
 internal class DefaultProjectSnapshotManagerAccessor : ProjectSnapshotManagerAccessor, IDisposable
 {
@@ -129,9 +61,6 @@ internal class DefaultProjectSnapshotManagerAccessor : ProjectSnapshotManagerAcc
         _optionsMonitor = optionsMonitor;
         _workspaceFactory = workspaceFactory;
         _notifierService = notifier;
-        //_generatorSnapshotFactory = snapshotFactory;
-
-        //Microsoft.CodeAnalysis.CodeAnalysisEventSource.Log.Message("project accessor made. notifier is :" + (notifier is not null));
     }
 
     public override ProjectSnapshotManagerBase Instance
